step3: views into separate package
